rename: context, stories 파일명 수정
